feat: add abstract recruit screen
